pentakit/requirements.txt
# Core
click>=8.1.0
rich>=13.0.0
httpx>=0.27.0
asyncio>=3.4.3
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# DNS
dnspython>=2.6.0

# Network / MITM
scapy>=2.5.0

# SSL
cryptography>=42.0.0

# MongoDB
pymongo>=4.6.0

# Reports
jinja2>=3.1.0
weasyprint>=62.0

# Notifications
python-telegram-bot>=21.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
tqdm>=4.66.0

# Scanning
python-nmap>=0.7.1

# Metadata
exifread>=3.0.0
PyPDF2>=3.0.0

# Async
aiohttp>=3.9.0
aiofiles>=23.0.0
